Naming convetion to delimiter separated
