+++ server/app/silc_api/models/accounting/asset_type_model.js
