+++ server/app/silc_api/routes/admin_router.js
