+++ server/app/silc_api/routes/silc/silc_groups_router.js
